feat: add cell_in in functionBC (#121)
- modify template parameter of a Cell
- add the cell inside the domain in the method which computes the values
in FunctionBC

+++ demos/FiniteVolume/level_set_from_scratch.cpp
@@ -334,7 +334,7 @@ bool update_mesh(Field& f, Field_u& u, Tag& tag)
     samurai::for_each_interval(mesh[SimpleID::cells],
                                [&](std::size_t level, const auto& interval, const auto& index_yz)
                                {
-                                   auto itag = static_cast<std::size_t>(interval.start + interval.index);
+                                   auto itag = interval.start + interval.index;
                                    for (int i = interval.start; i < interval.end; ++i)
                                    {
                                        if (tag[itag] & static_cast<int>(samurai::CellFlag::refine))
